separate out ToggleIcon and ToggleItem
